Refactor dicom_to_nifti tab
