+++ fsleyes_plugin_shimming_toolbox/st_plugin.py
@@ -22,15 +22,15 @@ import logging
 import os
 import wx
 
-from fsleyes_plugin_shimming_toolbox import __ST_DIR__, __CURR_DIR__
+from fsleyes_plugin_shimming_toolbox import __CURR_DIR__
 from fsleyes_plugin_shimming_toolbox.tabs.tab import Tab
 from fsleyes_plugin_shimming_toolbox.tabs.b0shim_tab import B0ShimTab
 from fsleyes_plugin_shimming_toolbox.tabs.b1shim_tab import B1ShimTab
+from fsleyes_plugin_shimming_toolbox.tabs.dicom_to_nifti_tab import DicomToNiftiTab
 from fsleyes_plugin_shimming_toolbox.components.dropdown_component import DropdownComponent
 from fsleyes_plugin_shimming_toolbox.components.run_component import RunComponent
 from fsleyes_plugin_shimming_toolbox.components.input_component import InputComponent
 
-from shimmingtoolbox.cli.dicom_to_nifti import dicom_to_nifti_cli
 from shimmingtoolbox.cli.mask import box, rect, threshold, sphere
 from shimmingtoolbox.cli.prepare_fieldmap import prepare_fieldmap_cli
 
@@ -455,47 +455,3 @@ class MaskTab(Tab):
         )
         sizer = run_component.sizer
         return sizer
-
-
-class DicomToNiftiTab(Tab):
-    def __init__(self, parent, title="Dicom to Nifti"):
-        description = "Convert DICOM files into NIfTI following the BIDS data structure"
-        super().__init__(parent, title, description)
-
-        self.sizer_run = self.create_sizer_run()
-        sizer = self.create_dicom_to_nifti_sizer()
-        self.sizer_run.Add(sizer, 0, wx.EXPAND)
-
-        self.parent_sizer = self.create_sizer()
-        self.SetSizer(self.parent_sizer)
-
-    def create_dicom_to_nifti_sizer(self):
-        path_output = os.path.join(__CURR_DIR__, "output_dicom_to_nifti")
-        input_text_box_metadata = [
-            {
-                "button_label": "Input Folder",
-                "button_function": "select_folder",
-                "name": "input",
-                "required": True
-            },
-            {
-                "button_label": "Subject Name",
-                "name": "subject",
-                "required": True
-            },
-            {
-                "button_label": "Config Path",
-                "button_function": "select_file",
-                "default_text": os.path.join(__ST_DIR__, "dcm2bids.json"),
-                "name": "config",
-            },
-            {
-                "button_label": "Output Folder",
-                "button_function": "select_folder",
-                "default_text": path_output,
-                "name": "output",
-            }
-        ]
-        component = InputComponent(self, input_text_box_metadata, cli=dicom_to_nifti_cli)
-        run_component = RunComponent(panel=self, list_components=[component], st_function="st_dicom_to_nifti")
-        return run_component.sizer
